Sample moved into the separate directory
