+++ runtime/sample/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "parity-runtime-example"
+version = "0.1.1"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+
+description = "Example for Tokio runtime wrapper"
+license = "GPL-3.0"
+homepage = "https://www.parity.io/"
+
+[dependencies]
+futures = "0.1"
+tokio = "0.1.22"
+parity-runtime = { path = ".." }
+
+[workspace]
